Separated batching.js into file per class it contained
