+++ src/index.js
@@ -71,7 +71,10 @@ export { VertexIterator } from './graphics/vertex-iterator.js';
 export * from './scene/constants.js';
 export { calculateNormals, calculateTangents, createBox, createCapsule, createCone, createCylinder, createMesh, createPlane, createSphere, createTorus } from './scene/procedural.js';
 export { BasicMaterial } from './scene/materials/basic-material.js';
-export { Batch, BatchGroup, BatchManager } from './scene/batching.js';
+export { Batch } from './scene/batching/batch.js';
+export { BatchGroup } from './scene/batching/batch-group.js';
+export { SkinBatchInstance } from './scene/batching/skin-batch-instance.js';
+export { BatchManager } from './scene/batching/batch-manager.js';
 export { Camera } from './scene/camera.js';
 export { DepthMaterial } from './scene/materials/depth-material.js';
 export { ForwardRenderer } from './scene/forward-renderer.js';
